requirements.txt
Flask==3.0.3
ffmpeg-python==0.2.0
